Add remainder of process loop
This commit finishes the process loop to:
- build the paragraphs
- find quotes out of the paragraphs which match the right length
- print the book and quote information

While finding the quote, the logic is simplified quite a bit from
the perl version by removing unnecessary checks.  Here we only need
to verify the paragraph starts and ends with a quotation mark, and
if the correct length.

Additionally, this version is storing all of the matched quotes
and selecting the last one to be found, where the perl version
is only storing the last match.  A future change will randomly
select a quote from the matches if more than one is found.