Refactor output and exit status
This commit makes a few general output changes:

- moves logging output to the log package instead of fmt

The logging output should be going to STDERR instead of STDOUT.
Using log instead of fmt does that, as well as handles exit status.

Until logging to a file is implemented, this demonstrates capturing
both:

    $ go run quote.go -m 220 -d > quote.out 2> error.out
    $ cat quote.out

    title: The Secret Sharer
    author: Joseph Conrad

    "I won't be there to see you go," I began with an effort. (...)

    $ cat error.out
    [debug] page_link: https://gutenberg.org/ebooks/220
    [debug] book_link: https://gutenberg.pglaf.org/2/2/220/220.txt
    [debug] parser is in head
    [debug] parser is in body
    [debug] parser is in footer
    [debug] title: The Secret Sharer
    [debug] author: Joseph Conrad
    [debug] paragraphs found: 356
    [debug] quote was found: "Hadn't the slightest idea. I am (...)
    [debug] quote was found: "Unless you manage to recover    (...)
    [debug] quote was found: "Bless my soul! Do you mean, sir (...)
    [debug] quote was found: "I won't be there to see you go, (...)

- updates all error exit status code 1 instead of 2

While exit status 2 is still correct to detect issue, exit status 1
(as far as I'm concerned) is a better understood generic status for
error.

- switch the final output to Printf instead of Print

Rather than concatenate the strings to print, the format specifier
is a "better practice" for printing strings with multiple variable
parts.