Fix download_error_count and add status check
The download_error_count variable was incorrectly defined within
the MAIN loop, so wasn't being incremented.

This commit also changes the way download errors are handled.
If the GET request returns an error, immediately exit with failure.
If the GET request returns anything but 200, then follow the
previous logic to increment download_error_count or exit.